backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.0.0
cryptography==41.0.7
python-dotenv==1.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.7
django-environ==0.10.0
pyyaml==6.0.1
redis==4.6.0
djangorestframework-simplejwt==5.3.0